More rename threedi_tools=>threedi_plugin
