+++ tool_animation/test_tool_map_animator.py
@@ -1,4 +1,4 @@
-from ThreeDiToolbox.threedi_tools import CommandBox
+from ThreeDiToolbox.threedi_plugin import CommandBox
 from ThreeDiToolbox.tool_result_selection.models import TimeseriesDatasourceModel
 
 import mock
